moved ureg to tritium
