+++ libra_toolbox/tritium/__init__.py
@@ -1,1 +1,7 @@
+import pint
+
+ureg = pint.UnitRegistry()
+ureg.setup_matplotlib()
+ureg.define("neutron = 1 * particle = n")
+
 from . import model
